vision/services/vision-router/requirements.txt
fastapi>=0.116.0
uvicorn==0.30.6
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.4.0
tritonclient[http]>=2.47.0
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.11.0
starlette>=0.49.1
